Convert sleighState back to an environment to see if this fixes the problem with eachWorker() calls hanging after the 1st.   This seems to be a problem in that the master and the workers get out-of-synch on which barrier to use.  It would probably be best to modify the NWS protocol so that the barrier name is explicitly provided instead of relying on the workers and the master staying properly synchronized.
Use named functions rather than anonymous functions for eachWorker to make debugging easier.

Always use the sleighPending code to handle retrieving results, so that only
one code path is present, also to ease debugging and maintenance.

Use getElement instead of '[[' to extract elements for eachElement so that we can specify to use rows/cols or cells.


git-svn-id: svn://svn.r-forge.r-project.org/svnroot/r-gregmisc@778 45178186-c70b-4eec-b05e-9f6c5a094bf4